refactor: move configure credential to access aws handler
Move AWS credential configuration from `guard configure credential` to
`guard access aws` as a separate chariot handler group. This keeps
configure simple and prepares for future GCP/Azure support under the
access group. Uses @cli_handler for clean SDK injection instead of
manual context traversal.
