+++ praetorian_cli/main.py
@@ -1,5 +1,6 @@
 import click
 
+import praetorian_cli.handlers.access
 import praetorian_cli.handlers.add
 import praetorian_cli.handlers.aegis
 import praetorian_cli.handlers.agent
